Fix handling of face formats when some unused values may be specified
This resolves incorrectly assigning a normal to the Quad object in
quad.obj. This occurs if we've specified some data in the file but don't
use it for the faces being specified. Since the VertexIndices were a
usize we could only check if the tmp normals/texcoords vecs where empty
to see if we should be assigning values for them. This is wrong in the
case below where we specify a normal but don't use it in the first
object

v 0 1 0
v 0 0 0
v 1 0 0
vn 0 0 1
o No_Normal_for_me
f 1 2 3
o I_use_the_normal
f 1//1 2//1 3//1