Support loading files from buffers
This makes the `load_obj_buf` and `load_mtl_buf` functions public.

In order to handle the case where an in-memory `.obj` file references a
`.mtl` file, we add the concept of a "material loader".

This lets downstream users choose where to load a given material from
given a path.

The `load_obj` method has its own material loader which opens files from
the same directory as the obj file.